requirements.txt
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
pandas==1.3.3
polars==0.8.17
pytz==2021.1
seaborn==0.11.2
pyyaml==5.4.1
scikit-learn==0.24.2
